attendance_system_flask/requirements.txt
Flask
python-dotenv
supabase
gunicorn
bcrypt
